Full Station generalization

+++ Assets/Scripts/Non Mono/Events/HUD/HUDEvents.cs
@@ -2,5 +2,5 @@ using System;
 
 public class HUDEvents
 {
-    public static Action<GatheringType, int> OnResourceInventoryUpdate;
+    public static Action<GatheringType, float> OnResourceInventoryUpdate;
 }
